nixos/snipe-it: fix setup service
After the first run, when the non-existing file at
"$invalid_barcode_location" is first created, the script returned with
exit code 0, after `cp` was successful.
On a consecutive run, when the file already existed and the test
condition was therefore not satisfied, the exit code of this setup script
was 1, which caused this service to fail and consequently kept the main
service from starting.